Move platforms out of platform/
