+++ bottery/bottery.py
@@ -1,5 +1,6 @@
 import asyncio
 import importlib
+import logging
 from datetime import datetime
 
 import aiohttp.web
@@ -8,6 +9,7 @@ import click
 import bottery
 from bottery.conf import settings
 from bottery.log import Spinner
+
 
 
 class Bottery:
@@ -63,11 +65,7 @@ class Bottery:
             platform['OPTIONS'].update(global_options)
             platform['OPTIONS']['engine_name'] = engine_name
 
-            try:
-                mod = importlib.import_module(platform['ENGINE'])
-            except ImportError:
-                # TODO: log
-                continue
+            mod = importlib.import_module(platform['ENGINE'])
 
             with Spinner('Configuring %s' % engine_name.title()):
                 engine = mod.engine(**platform['OPTIONS'])
