--- bottery/platform/messenger/__init__.py
@@ -1,5 +1,0 @@
-from bottery.platform.messenger.api import MessengerAPI  # noqa
-from bottery.platform.messenger.engine import MessengerEngine
-
-
-engine = MessengerEngine
