--- bottery/platform/telegram/__init__.py
@@ -1,6 +1,0 @@
-from bottery.platform.telegram.api import TelegramAPI  # noqa
-from bottery.platform.telegram.engine import TelegramEngine as engine  # noqa
-from bottery.platform.telegram.widgets import (  # noqa
-    Keyboard as keyboard,
-    Reply as reply,
-)
