+++ bottery/telegram/__init__.py
@@ -1,0 +1,6 @@
+from bottery.telegram.api import TelegramAPI  # noqa
+from bottery.telegram.engine import TelegramEngine as engine  # noqa
+from bottery.telegram.widgets import (  # noqa
+    Keyboard as keyboard,
+    Reply as reply,
+)
