+++ tests/test_platform_telegram.py
@@ -1,9 +1,9 @@
 import pytest
 
 from bottery.message import Message
-from bottery.platform.telegram import reply
-from bottery.platform.telegram.engine import (TelegramChat, TelegramEngine,
-                                              TelegramUser)
+from bottery.telegram import reply
+from bottery.telegram.engine import (TelegramChat, TelegramEngine,
+                                     TelegramUser)
 
 
 @pytest.fixture
